Add behavior refactoring in simple braitenberg env
